requirements.txt
google
pandas
numpy
scipy
pillow
newspaper3k
fake_useragent
openpyxl
httplib2